defined motorcontroller, cleaned up naming
