+++ MotorController/SOTA_MotorController.java
@@ -6,46 +6,130 @@ import SOTAlib.Encoder.SOTA_Encoder;
 public interface SOTA_MotorController extends MotorController {
 
     /**
-     * @param neutralOperation
+     * @param neutralOperation desired neutral state
      */
     void setNeutralOperation(NeutralOperation neutralOperation);
 
+    /**
+     * Returns whole encoder object to access methods from it
+     * example: SOTA_MotorController.getEncoder().getAbsolutePosition()
+     * 
+     * @return SOTA_Encoder object
+     */
     SOTA_Encoder getEncoder();
 
+    /**
+     * @return encoder velocity in RPM
+     */
     double getEncoderVelocity();
 
+    /**
+     * @return encoder position in rotations
+     */
     double getEncoderPosition();
 
-    double getNativeEncoderVelocity();
+    /**
+     * This method accesses the encoder built in to the motor, not the composite
+     * SOTA_MotorController SOTA_Encoder combo
+     * 
+     * @return integrated encoder velocity in RPM
+     */
+    double getIntegratedEncoderVelocity();
 
-    double getNativeEncoderPosition();
+    /**
+     * This method accesses the encoder built in to the motor, not the composite
+     * SOTA_MotorController SOTA_Encoder combo
+     * 
+     * @return integrated encoder position in Rotations
+     */
+    double getIntegratedEncoderPosition();
 
-    double getNativeCountsPerRevolution();
-
+    /**
+     * Resets position of integrated encoder, not the composite SOTA_MotorController
+     * SOTA_Encoder combo
+     */
     void resetNativeEncoder();
 
+    /**
+     * 
+     * @return Motor Temp in C
+     */
     double getMotorTemperature();
 
+    /**
+     * 
+     * @return Motor current in Amps
+     */
     double getMotorCurrent();
 
+    /**
+     * Sets the maximum amount of current a motor is allowed to draw
+     * 
+     * @param amps maximum amps
+     */
     void setCurrentLimit(int amps);
 
-    MotorLimits getLimits();
+    /**
+     * Returns the MotorPositionLimits object that contains the upper and lower
+     * bound of the
+     * encoder range the motor is allowed to operate in
+     * 
+     * @return MotorLimits object
+     */
+    MotorPositionLimits getLimits();
 
-    void setLimits(MotorLimits limits);
+    /**
+     * Sets the upper and lower bound of the encoder range the motor is allowed to
+     * operate in
+     * 
+     * @param limits MotorLimits object with new limits
+     */
+    void setPositionLimits(MotorPositionLimits limits);
 
+    /**
+     * 
+     * @return the lower bound of the encoder range the motor is allowed to operate
+     *         in
+     */
     double getLowerLimit();
 
+    /**
+     * 
+     * @return the upper bound of the encoder range the motor is allowed to operate
+     *         in
+     */
     double getUpperLimit();
 
+    /**
+     * 
+     * @return if the encoder is at the upper limit defined from the MotorLimits
+     *         object
+     */
     boolean atUpperLimit();
 
+    /**
+     * 
+     * @return if the encoder is at the lower limit defined from the MotorLimits
+     *         object
+     */
     boolean atLowerLimit();
 
+    /**
+     * 
+     * @return State of the interaction between the Encoder Position and the Range
+     *         of the Limits, see EncoderLimitStates Enum for more information
+     */
+    MotorPositionLimitStates getLimitState();
+
+    /**
+     * Resets the encoder to 0
+     */
     void resetEncoder();
 
-    boolean getNeutralOperation();
-
-    void setNeutralOperation();
+    /**
+     * 
+     * @return Current Neutral Operation from NeutralOperation Enum
+     */
+    NeutralOperation getNeutralOperation();
 
 }
